pyproject.toml
[project]
name = "bub-contrib"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bub",
    "bub-codex",
    "bub-discord",
    "bub-dingtalk",
    "bub-github-copilot",
    "bub-kimi",
    "bub-mcp",
    "bub-qq",
    "bub-schedule",
    "bub-searxng-search",
    "bub-tapestore-sqlalchemy",
    "bub-tapestore-redis",
    "bub-tapestore-sqlite",
    "bub-tg-feed",
    "bub-web-search",
    "bub-feishu",
    "bub-session-prompt",
    "bub-wechat",
    "bub-wecom",
    "tape-dataset-opendal",
]

[tool.uv.workspace]
members = ["packages/*"]

[tool.uv.sources]
bub = { git = "https://github.com/bubbuild/bub.git" }
bub-tg-feed = { workspace = true }
bub-codex = { workspace = true }
bub-discord = { workspace = true }
bub-dingtalk = { workspace = true }
bub-github-copilot = { workspace = true }
bub-kimi = { workspace = true }
bub-mcp = { workspace = true }
bub-qq = { workspace = true }
bub-schedule = { workspace = true }
bub-searxng-search = { workspace = true }
bub-tapestore-sqlalchemy = { workspace = true }
bub-tapestore-redis = { workspace = true }
bub-tapestore-sqlite = { workspace = true }
bub-web-search = { workspace = true }
bub-feishu = { workspace = true }
bub-session-prompt = { workspace = true }
bub-wechat = { workspace = true }
bub-wecom = { workspace = true }
tape-dataset-opendal = { workspace = true }

[dependency-groups]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "fakeredis>=2.20.0",
]

[tool.pytest.ini_options]
addopts = "-ra --import-mode=importlib"
